Add support for z/OS (IBM) system
Introduction
------------

Currently, NppFTP works well for all Unix-like systems.
NppFTP only works if the server uses "/" as separator.
z/OS systems don't work like Unix and thus can't work with NppFTP.
This patch may seem useless because so few people use old z/OS systems
but it's the contrary, this patch is a major improvement for NppFTP.
z/OS systems are used in major company, insurance, government (to compute
your taxes for example).

I work for Linagora and our customer (French ministry) needs a great
tool like Notepad++ to work on their z/OS system.
The consequences of this patch may be a major adoption of Notepad++ in
the biggest infrastructures which imply major support and so
big improvement of Notepad++.

This is great for free software!

Now that I have convinced you ;-), let's talk about techniques.

Note
----

I know you probably can't test this patch so you have to trust me on this.

Context
-------

When accessing a 'folder' (it's not really a folder but it looks like),
you must use single quote around the path ('MYROOT.MYFOLDER').
It's strange but z/OS is an old grandfather of computing.
Moreover, z/OS doesn't use "/" but "." as separator and nested folders
are forbidden, hierarchy tree can be only of one level (neat!).

The task here is to keep the plugin as pure as possible, I don't want
to change the data model or the code, I want to be as less intrusive
as possible to avoid side effects.

How it works
------------

Firstly, I have added a new property in profile to say to NppFTP:
This profile is a z/OS system.

To do that, I have added a `m_serverType` in `FTPProfile.cpp` and
updated the `ProfilesDialog.cpp` to work with this new property.

Next, I have created a `FTPClientWrapperZOS` extending `FTPClientWrapperSSL`.
The main goal of this wrapper is to convert Unix-path to z/OS-path.
Don't forget, I try to be as less intrusive as possible.
This wrapper is initialized in `FTPProfile::CreateWrapper`
if `m_serverType` is z/OS.

All FTP requests are now using this wrapper and it works great.

Nevertheless, two problems still remain:

1 - "/" is assumed as directory separator
2 - When listing folder, z/OS format is different of Unix one

To resolve 1:

In `FileObject` constructor, if "/" is not found, we set the name
equals to the full path.
In `FTPCache::IsPathPrefixLocal`, if `localpath` is not starting
by "/" we set `postfix` equals to `localpath`.

To resolve 2:

Everything is in `ftp_c.cpp`.
In `GetDirInfoPASV` (z/OS only support passive connection), we detect
if the response is a z/OS response and we parse it with the
corresponding method.

There are two ways of parsing z/OS output:

- In root: `GetInfoInZOSFormat`
- In folder: `GetInfoInZOSFormatContainer`

Like you can see, there is a `parseMigrated` function.

Migrated is a specific file which needs to be accessed a first time
to retrieve informations about it. So the first time we access
a migrated file, we set default values since we don't know real values.

Conclusion
----------

I wait for your notes about this patch. You should know that
I have no access to a z/OS system anymore so if I update the code,
I can't test it.

Regards,
Jean-Sebastien Bevilacqua