updated to libssh-0-9-1 (#266)
, see https://www.libssh.org/2019/10/25/libssh-0-9-1/

The libssh team is proud to announce the first bugfix release of libssh as version 0.9.1. It fixes bugs we discovered since the last release. Thanks to all contributors!

There are two things new and this is that we use X22919 and Ed25519 from OpenSSL if it provides them.

If you are new to libssh you should read our tutorial how to get started. Please join our mailing list or visit our IRC channel if you have questions.

You can download libssh here.
Changelog

    Added support for Ed25519 via OpenSSL
    Added support for X25519 via OpenSSL
    Added support for localuser in Match keyword
    Fixed Match keyword to be case sensitive
    Fixed compilation with LibreSSL
    Fixed error report of channel open (T75)
    Fixed sftp documentation (T137)
    Fixed known_hosts parsing (T156)
    Fixed build issue with MinGW (T157)
    Fixed build with gcc 9 (T164)
    Fixed deprecation issues (T165)
    Fixed known_hosts directory creation (T166) Fixed some issues with the connector API